requirements.txt
watchdog
python-telegram-bot
python-dotenv